requirements.txt
claude-agent-sdk==0.1.58
PyGithub==2.6.0
requests==2.32.3
python-dotenv==1.1.0
